Display Tweet form modally
Re-arrange the `tweets#index` template's [`<turbo-frame>`
element][turbo-frame] to be a descendant of a [`<dialog>`
element][dialog] that's a direct descendant of the [`<body>`
element][body] and a sibling of the page's [`<main>` element][main].

Ideally, `<turbo-frame>` elements would dispatch [`turbo:`-prefixed
events][events] that reflect their progress throughout their individual
life-cycle. In the meantime, this commit utilizes [MutationObserver][]
to monitor `<turbo-frame>` elements that are descendants of `<dialog>`
elements, in an effort to display the `<dialog>` modally when the
initiated request completes.

[turbo-frame]: https://turbo.hotwire.dev/reference/frames
[dialog]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/dialog
[body]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/body
[main]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/main
[events]: https://turbo.hotwire.dev/reference/events
[MutationObserver]: https://developer.mozilla.org/en-US/docs/Web/API/MutationObserver